+++ db/migrate/20151105160658_drop_subject_schedules_table.rb
@@ -1,0 +1,14 @@
+class DropSubjectSchedulesTable < ActiveRecord::Migration
+  def up
+    drop_table :subject_schedules
+  end
+
+  def down
+    create_table :subject_schedules do |t|
+      t.integer :subject_id
+      t.integer :schedule_id
+      t.date :initial_due_date
+      t.timestamps
+    end
+  end
+end