+++ db/migrate/20151105161632_remove_schedule_columns_from_sheets.rb
@@ -1,0 +1,15 @@
+class RemoveScheduleColumnsFromSheets < ActiveRecord::Migration
+  def up
+    remove_index :sheets, :subject_schedule_id
+    remove_column :sheets, :subject_schedule_id
+    remove_index :sheets, :event_id
+    remove_column :sheets, :event_id
+  end
+
+  def down
+    add_column :sheets, :event_id, :integer
+    add_index :sheets, :event_id
+    add_column :sheets, :subject_schedule_id, :integer
+    add_index :sheets, :subject_schedule_id
+  end
+end